django_piscine/day05/requirements.txt
asgiref==3.2.3
Django==1.9
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
